Resolved requested changes.
